+++ crates/swc/tests/errors/lints/no-empty-function/default/ts/input.ts
@@ -1,0 +1,20 @@
+class A {
+  private constructor() {}
+
+  private method1() {}
+
+  protected method2() {}
+
+  public method3() {}
+
+  // max markers count
+  public async *method4() {}
+}
+
+class B {
+  constructor(params: number) {
+    class Nested {
+      method() {}
+    }
+  }
+}